+++ lib/src/features/records/ui/record_page/view/custom/record_view.dart
@@ -1,0 +1,19 @@
+part of '../record_page.dart';
+
+class _RecordView extends StatefulWidget {
+  const _RecordView({super.key});
+
+  @override
+  State<_RecordView> createState() => _RecordViewState();
+}
+
+class _RecordViewState extends State<_RecordView> {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: const [
+        _RecordGraph(),
+      ],
+    );
+  }
+}
