+++ lib/src/global/ui/navigation/app_router.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:go_router_flow/go_router_flow.dart';
 import 'package:nobook/src/features/features_barrel.dart';
-import 'package:nobook/src/features/records/domain/view/record_screen.dart';
 import 'package:nobook/src/features/records/ui/record_page/view/record_page.dart';
 import 'package:nobook/src/global/global_barrel.dart';
 
@@ -39,7 +38,6 @@ final GoRouter _router = GoRouter(
             return const DashBoardScreen();
           },
         ),
-        
         GoRoute(
           path: AppRoute.note.path,
           name: AppRoute.note.name,
@@ -58,21 +56,13 @@ final GoRouter _router = GoRouter(
             ),
           ],
         ),
-          GoRoute(
-          path: AppRoute.record.path,
-          name: AppRoute.record.name,
-          builder: (context, state) {
-            return const RecordScreen();
-          },
-        ),
-         GoRoute(
+        GoRoute(
           path: AppRoute.record.path,
           name: AppRoute.record.name,
           builder: (context, state) {
             return const RecordPage();
           },
         ),
-        // routes:
       ],
     ),
   ],
